FIX computeDayDelta now works correctly
computeDayDelta didn't calculate the total number of days between two given dates.
For Example: difference between 1970-01-01 and 2017-12-06 was 5 because the difference
between days (01 to 06) is 5. Reason for that was the use of Period.between.
The correct behaviour can be achieved with using ChronoUnit.DAYS.between
See: https://docs.oracle.com/javase/tutorial/datetime/iso/period.html